+++ src/main/java/com/Perfulandia_2025/Perfulandia_2025/requestDTO/ItemPedidoRequestDTO.java
@@ -2,14 +2,16 @@ package com.Perfulandia_2025.Perfulandia_2025.requestDTO;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ItemPedidoRequestDTO {
 
     private String descripcion;
-    private int cantidadSolicitada;
-    private int precioUnitario;
+    private Integer cantidadSolicitada;
+    private BigDecimal precioUnitario;
     private String codigoProducto;
 
 }
