+++ src/main/java/com/Perfulandia_2025/Perfulandia_2025/responseDTO/ItemPedidoResponseDTO.java
@@ -3,6 +3,8 @@ package com.Perfulandia_2025.Perfulandia_2025.responseDTO;
 import com.Perfulandia_2025.Perfulandia_2025.modelo.PedidoReabastecimiento;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,8 +13,8 @@ public class ItemPedidoResponseDTO {
     private Long id;
     private PedidoReabastecimiento pedidoReabastecimiento;
     private String descripcion;
-    private int cantidadSolicitada;
-    private int precioUnitario;
+    private Integer cantidadSolicitada;
+    private BigDecimal precioUnitario;
     private String codigoProducto;
 
 }
